refactor: extract line constructions to line/LineConstructions.ts
Moves all 21 line construction classes from Constructions.ts to a new
line/LineConstructions.ts. Includes the wrapPointAsLine helper and all
5 LinePerpendicular variants that extend point perpendicular classes.

Constructions.ts: 1364 → 1057 lines.

Build passes, 98 tests pass.
